+++ packages/app-builder/src/components/CaseManagerV2/AiReview/AiReviewPanel.tsx
@@ -10,6 +10,7 @@ import toast from 'react-hot-toast';
 import { useTranslation } from 'react-i18next';
 import { Button, Markdown, Tag, Typo } from 'ui-design-system';
 import { Icon } from 'ui-icons';
+import { AiReviewStatusMessage } from './AiReviewStatusMessage';
 
 type AiReviewPanelProps = {
   caseId: string;
@@ -101,7 +102,7 @@ function AiReviewPanelContent({ caseId, canManuallyReview, onOpenChange, reviews
             ) : review?.review ? (
               <Markdown>{review.review.output}</Markdown>
             ) : review ? (
-              <NoContentBody status={selectedListItem.status} />
+              <AiReviewStatusMessage status={selectedListItem.status} className="h-full justify-center" />
             ) : null}
           </div>
           {review?.review && !review.review.ok ? <SanityCheckWarning message={review.review.sanityCheck} /> : null}
@@ -110,31 +111,6 @@ function AiReviewPanelContent({ caseId, canManuallyReview, onOpenChange, reviews
       </Panel.Content>
     </Panel.Container>
   );
-}
-
-function NoContentBody({ status }: { status: AiCaseReviewStatus }) {
-  const { t } = useTranslation(['cases']);
-  if (status === 'pending') {
-    return (
-      <div className="flex h-full items-center justify-center gap-xs text-small text-grey-secondary">
-        <Icon icon="spinner" className="size-4 animate-spin" />
-        <span>{t('cases:case.ai_reviews.generating')}</span>
-      </div>
-    );
-  }
-  if (status === 'failed' || status === 'insufficient_funds') {
-    return (
-      <div className="flex h-full items-center justify-center gap-xs text-small text-red-primary">
-        <Icon icon="warning" className="size-4 shrink-0" />
-        <span>
-          {status === 'insufficient_funds'
-            ? t('cases:case.ai_reviews.insufficient_funds')
-            : t('cases:case.ai_reviews.failed')}
-        </span>
-      </div>
-    );
-  }
-  return null;
 }
 
 function SanityCheckWarning({ message }: { message: string }) {
